backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
alembic==1.13.1

# AI/ML
openai==1.54.3
langchain==0.1.4
langchain-openai==0.0.5
pgvector==0.2.4

# Document Processing
PyPDF2==3.0.1
pdf2image==1.17.0
Pillow==10.2.0
python-magic==0.4.27

# AWS
boto3==1.34.34

# Utilities
aiofiles==23.2.1
